refactor: split fat CLI files, share test chdir helper, harden CI/build
Bundles four small refactors that share no behavior change.

- internal/cli: split root.go command builders into one file per command
  (sync.go, validate.go, list.go) and split import_claude.go (308 LOC)
  into rules/skills/hooks files. Each file now under 120 LOC with one
  concern.
- internal/testutil: extract Chdir/TempCwd helpers used in 14 tests
  across adapters, cli, and integration. Removes 11 duplicated chdir
  blocks; adds t.Helper() consistently.
- internal/adapters/codex: convert table-driven cases in TestRouteDir
  and TestEmit_NestedByGlobs to t.Run subtests for better failure
  reports.
- ci/build: add coverage upload (codecov, ubuntu-only), -trimpath
  -ldflags="-s -w" to Make build, lint/fmt/vet/cover Make targets,
  Dependabot weekly for gomod and github-actions, SECURITY.md.

No behavior change. All tests pass.
