+++ SECURITY.md
@@ -1,0 +1,17 @@
+# Security Policy
+
+## Supported versions
+
+Only the latest minor release receives security fixes.
+
+## Reporting a vulnerability
+
+Report vulnerabilities through GitHub private security advisories:
+https://github.com/Chemaclass/agnostic-ai/security/advisories/new
+
+Do not open public issues for security reports.
+
+## Disclosure window
+
+We aim to publish a fix or mitigation within 90 days of a confirmed report.
+After that window, the report may be disclosed publicly.
