+++ internal/cli/root.go
@@ -2,12 +2,8 @@
 package cli
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/spf13/cobra"
 
-	"github.com/chemaclass/agnostic-ai/internal/adapters"
 	"github.com/chemaclass/agnostic-ai/internal/config"
 	"github.com/chemaclass/agnostic-ai/internal/spec"
 )
@@ -29,103 +25,6 @@ func NewRootCmd(version string) *cobra.Command {
 		newDoctorCmd(),
 	)
 	return root
-}
-
-func newSyncCmd() *cobra.Command {
-	var targets []string
-	var dryRun, check bool
-
-	cmd := &cobra.Command{
-		Use:   "sync",
-		Short: "Emit per-target configs from agnostic specs.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if check {
-				reports, err := collectDrift(targets)
-				if err != nil {
-					return err
-				}
-				if printDrift(reports) {
-					return fmt.Errorf("drift detected")
-				}
-				return nil
-			}
-			cfg, b, err := loadProject(".")
-			if err != nil {
-				return err
-			}
-			if len(targets) == 0 {
-				targets = cfg.Targets
-			}
-			for _, t := range targets {
-				adapter, ok := adapters.Get(t)
-				if !ok {
-					fmt.Fprintf(os.Stderr, "! unknown target: %s\n", t)
-					continue
-				}
-				fmt.Printf("→ emit %s\n", t)
-				if err := adapter.Emit(b, cfg, dryRun); err != nil {
-					return fmt.Errorf("%s: %w", t, err)
-				}
-			}
-			return nil
-		},
-	}
-	cmd.Flags().StringSliceVarP(&targets, "target", "t", nil, "Targets to emit (default: all in config)")
-	cmd.Flags().BoolVar(&dryRun, "dry-run", false, "Print outputs instead of writing")
-	cmd.Flags().BoolVar(&check, "check", false, "Compare emitted output to disk; non-zero exit on drift")
-	return cmd
-}
-
-func newValidateCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "validate",
-		Short: "Validate agnostic specs.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			_, b, err := loadProject(".")
-			if err != nil {
-				return err
-			}
-			fmt.Printf("loaded %d entries. ok.\n", len(b.All()))
-			return nil
-		},
-	}
-}
-
-func newListCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "list",
-		Short: "List loaded specs.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			_, b, err := loadProject(".")
-			if err != nil {
-				return err
-			}
-			for _, e := range b.All() {
-				fmt.Printf("%s\t%s\n", e.Kind, e.Name)
-			}
-			return nil
-		},
-	}
-}
-
-func newInitCmd() *cobra.Command {
-	var from string
-	cmd := &cobra.Command{
-		Use:   "init",
-		Short: "Scaffold an agnostic-ai project in the current directory.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			switch from {
-			case "":
-				return scaffold(".")
-			case "claude":
-				return importFromClaude(".")
-			default:
-				return fmt.Errorf("unknown source for --from: %q (supported: claude)", from)
-			}
-		},
-	}
-	cmd.Flags().StringVar(&from, "from", "", "Import existing config from a source (supported: claude)")
-	return cmd
 }
 
 // loadProject loads config and bundle from root.
